Catch ILS exceptions in HoldsController.
Moves the existing OnDispatch handler that catches ILS exceptions from MyResearchController to a trait and uses it in MyResearchController and HoldsController.
